Add menu to rename and delete homes
This menu is triggered upon right clicking the bed elements. Alongside a
"Go Back" button is are "Rename" and "Delete" buttons. The "Rename"
button opens up a new anvil menu to rename the bed, while the delete
removes the bed from the menu and kicks the player back to the selection
menu.