Set home limit based on Vault metadata
By making use of the "Chat" metadata that Vault provides, certain values
such as setting a home limit can be done externally as a group or for an
individual player through a permission plugin. The method in doing so is
usually to do with chat prefixes and colours, but doesn't provide any
restrictions in using it for any other primitive values.