Add new check if player tries to teleport to locked warp
The menu will also auto refresh if the player tries to teleporting to a
locked warp that changes to an unlocked state while they have the menu
still open. Otherwise if the menu opened while the warp was unlocked but
was locked after, it will output a message telling them it's been
locked now.